refactor: TOF高度估计全链路改为float并优化速度融合平滑性 并且飞行了一次

+++ project/code/Estimation/Height_Est/TOF_data.h
@@ -17,15 +17,15 @@ extern "C" {
 #endif
 
 /* 融合后TOF高度，单位：mm，无效时为VL53L1X_VALID_RANGE_MAX */
-extern uint16 g_tof_fused_height_mm;
+extern float g_tof_fused_height_mm;
 /* TOF通道1显示高度（中心高度估计）各通道中心高度估计（几何修正+零偏后），单位：mm，无效时为VL53L1X_INVALID_DISTANCE_MM */
-extern uint16 g_tof1_height_mm;
+extern float g_tof1_height_mm;
 /* TOF通道2显示高度（中心高度估计）各通道中心高度估计（几何修正+零偏后），单位：mm，无效时为VL53L1X_INVALID_DISTANCE_MM */
-extern uint16 g_tof2_height_mm;
+extern float g_tof2_height_mm;
 /* TOF通道3显示高度（中心高度估计）各通道中心高度估计（几何修正+零偏后），单位：mm，无效时为VL53L1X_INVALID_DISTANCE_MM */
-extern uint16 g_tof3_height_mm;
+extern float g_tof3_height_mm;
 /* TOF通道4显示高度（中心高度估计）各通道中心高度估计（几何修正+零偏后），单位：mm，无效时为VL53L1X_INVALID_DISTANCE_MM */
-extern uint16 g_tof4_height_mm;
+extern float g_tof4_height_mm;
 /* 融合高度有效标志，1=有效，0=无效  */ 
 extern uint8 g_tof_fused_valid;
 /* 通道1原始有效标志，1=有效，0=无效 */
@@ -45,10 +45,10 @@ extern uint8 g_tof_fused_source;
 /* 校准通过标志：1=通道间偏差在允许范围内，0=校准失败或偏差过大 */
 extern uint8 g_tof_calibration_ok;
 extern float g_tof_fused_vz_mps;
-extern uint16 g_tof_measure_height_mm;
+extern float g_tof_measure_height_mm;
 extern uint8 g_tof_measure_class;
 extern uint8 g_tof_measure_mask;
-extern uint16 g_tof_shadow_height_mm;
+extern float g_tof_shadow_height_mm;
 extern float g_tof_shadow_vz_mps;
 extern uint8 g_tof_shadow_valid;
